requirements.txt
python-telegram-bot==13.15
ldap3==2.9.1
python-dotenv==1.0.0
pycryptodome==3.20.0
urllib3>=1.26.0
six>=1.16.0
